requirements.txt
pandas==1.5.3
numpy==1.23.5
scikit-learn==1.1.3
matplotlib==3.7.1
seaborn==0.12.2
Pillow==9.5.0
torch==1.13.1
torchvision==0.14.1
pytorch-lightning==2.1.0
mlflow==2.1.0
scikit-image==0.21.0
tqdm==4.65.0